Update checkbox.css
As far I can tell, the z-index does not effect the rendering of the checkbox.
I stumbled upon problems where other third party elements (such as React-dropdown) would render under the checkbox. This has not been a problem with any other component.

Is there a reason it has the z-index in the first place?